Deduplicate function get_final_indices_from_shared_bitmap
